mod_1_chat/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version

functions-framework==3.*
google-cloud-logging==3.0.0

